enabled experimental support for atmega

+++ src/logger.hpp
@@ -95,7 +95,7 @@
   #define YAML_PATHNAME _pathToFileName
   static const char * _pathToFileName(const char * path)
   {
-    size_t i = 0, pos = 0;
+    int i = 0, pos = 0;
     char * p = (char *)path;
     while(*p){
       i++;
@@ -106,6 +106,8 @@
     }
     return path+pos;
   }
+  #include <stdarg.h>
+  #include <stdio.h>
 #endif
 
 #if !defined LOG_PRINTF
